Updating readme.md to include jx.lib download
While trying to build the project, I hit the error "Cannot find jx.lib". It was unclear to me initially that I had to install that lib in the external/ path before building the project. So, having the instruction to use the powershell script helped download the lib and build the project.